test(hooks): provision in-memory localStorage shim for jsdom suites
Node 22+ ships a built-in global `localStorage` that, without
`--localstorage-file`, exposes no working `Storage` (`.clear`/`.setItem`
absent) and shadows jsdom's `window.localStorage` (they are the same ref).
A bare `localStorage.clear()` in `beforeEach` then throws — green on CI's
Node 20, red on Node 25 locally — failing 13 tests across
useAnalyzedResume + useDownloadPdf and tripping the pre-push verify hook.

Extract the existing MemoryStorage shim (from useModelSelection.test.ts)
into a shared `__test-utils__/memory-storage.ts` helper and install a
fresh per-test global in the affected suites, so persisted-state tests
run correctly regardless of the runtime's native localStorage.

Refs #398
Refs #396
